feat: BeanDefinition 추가
