+++ nextstep-di/src/main/java/nextstep/di/BeanDefinition.java
@@ -1,0 +1,11 @@
+package nextstep.di;
+
+public interface BeanDefinition {
+    Class<?> getType();
+
+    Class<?>[] getParameterTypes();
+
+    boolean isType(Class<?> type);
+
+    Object createBean(Object... params);
+}
